Move renderer specific notifications to renderer folder
