+++ desktop/packages/mullvad-vpn/src/renderer/lib/notifications/index.ts
@@ -1,0 +1,2 @@
+export * from './new-device';
+export * from './new-version';
